feat(accessibility): add setting content description #157
* Added support for a11y contentDescription attribute

If empty, defaults to original primary and secondary texts
concatenation.

* Added custom contentDescription to sample

Added only to showSideNavigationPrompt(), to limit PR changes

* Handle a11y double-tap event as onClick()

Using talkback, double-tap does not do anything. Users need to do the
back gesture to close prompt.
Passing the click event to the targetView might not be the best a11y
solution, but currently fixes the issue for my app!